提交更新，此次更新无重大版本变革

+++ DB数据库/MySQL数据库/src/B2条件查询.java
@@ -6,8 +6,8 @@
      CREATE TABLE t_emp(                                              语法格式：
         id INT(20) PRIMARY KEY AUTO_INCREMENT COMMENT '主键Id',           SELECT
         ename VARCHAR(255) COMMENT '员工姓名',                                  字段1, 字段2, ...     3
-        sal DOUBLE COMMENT '工资',                                        FROM
-        comm DOUBLE COMMENT '津贴',                                             表名                 1
+        sal DOUBLE(10, 2) COMMENT '工资',                                 FROM
+        comm DOUBLE(10, 2) COMMENT '津贴',                                      表名                 1
         dept_no VARCHAR(255) COMMENT '部门',                              WHERE
         job VARCHAR(255) COMMENT '工作岗位',                                    条件                  2
         province VARCHAR(255) COMMENT '省份'
